translateField(fieldTranslation, value2)
