+++ src/services/tagging/idTaggingScheme.ts
@@ -10,6 +10,7 @@ import { Field } from './types/Fields';
 import { DEBUG_ID_SCHEMA } from '../../config.mjs';
 import { FEATURED_KEYS } from './featuredKeys';
 import { deduplicate } from './utils';
+import { KeysTodo, keysTodo } from './keysTodo';
 
 const logMoreMatchingFields = (matchingFields: Field[], key: string) => {
   if (DEBUG_ID_SCHEMA && matchingFields.length > 1) {
@@ -56,7 +57,7 @@ const getUiField = (
 
 const matchFieldsFromPreset = (
   preset: Preset,
-  keysTodo: any,
+  keysTodo: KeysTodo,
   feature: Feature,
 ): UiField[] => {
   const fieldKeys = getFieldKeys(preset);
@@ -94,59 +95,6 @@ const matchRestToFields = (keysTodo: KeysTodo, feature: Feature): UiField[] =>
     }
     return getUiField(field, keysTodo, feature, key);
   });
-
-type KeysTodo = typeof keysTodo;
-const keysTodo = {
-  state: [] as string[],
-  init(feature: Feature) {
-    this.state = Object.keys(feature.tags);
-  },
-  resolveTags(tags: FeatureTags) {
-    Object.keys(tags).forEach((key) => this.remove(key));
-  },
-  has(key: string) {
-    return this.state.includes(key);
-  },
-  hasAny(keys: string[]) {
-    return keys?.some((key) => this.state.includes(key));
-  },
-  remove(key: string) {
-    const index = this.state.indexOf(key);
-    if (index > -1) {
-      this.state.splice(index, 1);
-    }
-  },
-  removeByRegexp(regexp: RegExp) {
-    this.state = this.state.filter((key: string) => !regexp.test(key));
-  },
-  resolveFields(fieldsArray: UiField[]) {
-    fieldsArray.forEach((field) => {
-      if (field?.field?.key) {
-        this.remove(field.field.key);
-      }
-      if (field?.field?.keys) {
-        field.field.keys.forEach((key) => this.remove(key));
-      }
-    });
-  },
-  mapOrSkip<T>(fn: (key: string) => T): NonNullable<T>[] {
-    const skippedFields = [];
-    const output = [];
-
-    while (this.state.length) {
-      const field = this.state.shift();
-      const result = fn(field); // this can remove items from this.state
-      if (result) {
-        output.push(result);
-      } else {
-        skippedFields.push(field);
-      }
-    }
-
-    this.state = skippedFields;
-    return output;
-  },
-};
 
 const getFeaturedTags = (feature: Feature) => {
   const { tags } = feature;
